backend/train_models.py
"""
SDARS AI Training Script
Generates a technically-rigorous synthetic dataset and trains the 
Multi-Modal Disaster Prediction models (Random Forest).
"""
import numpy as np
import pandas as pd
from sklearn.ensemble import RandomForestClassifier
import joblib
import os
import config

def generate_training_data(n_samples=2000):
    """
    Generates data that follows the physics of real disasters.
    Features: [temp, hum, wind, press, ndvi, ndwi, hotspots]
    """
    np.random.seed(42)
    
    # 1. Base random data
    temp = np.random.uniform(10, 50, n_samples)
    hum = np.random.uniform(5, 100, n_samples)
    wind = np.random.uniform(0, 100, n_samples)
    press = np.random.uniform(980, 1030, n_samples)
    ndvi = np.random.uniform(0, 1, n_samples)
    ndwi = np.random.uniform(-0.5, 0.5, n_samples)
    hotspots = np.random.uniform(0, 5, n_samples)
    
    X = pd.DataFrame({
        'temp': temp, 'hum': hum, 'wind': wind, 'press': press,
        'ndvi': ndvi, 'ndwi': ndwi, 'hotspots': hotspots
    })
    
    # 2. Logic-based Labels (Ground Truth)
    # Fire: Hot, Dry, Windy, Low NDVI, High Hotspots
    y_fire = ((temp > 35) & (hum < 25) & (hotspots > 1.5) | (temp > 42)).astype(int)
    
    # Flood: Rainy (Humid), Low pressure, High NDWI
    y_flood = ((hum > 85) & (ndwi > 0.2) & (press < 1005)).astype(int)
    
    # Cyclone: High Wind, Very Low Pressure, Very Humid
    y_cyclone = ((wind > 50) & (press < 995) & (hum > 80)).astype(int)
    
    return X, y_fire, y_flood, y_cyclone

def train_and_save():
    print("🧠 AI CORE: Initializing synthetic dataset generator...")
    X, y_fire, y_flood, y_cyclone = generate_training_data()
    
    models = {
        'fire': (y_fire, 'fire_risk_model.joblib'),
        'flood': (y_flood, 'flood_risk_model.joblib'),
        'cyclone': (y_cyclone, 'cyclone_risk_model.joblib')
    }
    
    for name, (y, filename) in models.items():
        print(f"🛰️  Training {name.upper()} multi-modal classifier...")
        model = RandomForestClassifier(n_estimators=100, random_state=42)
        model.fit(X, y)
        
        path = os.path.join(config.MODELS_DIR, filename)
        joblib.dump(model, path)
        print(f"✅ Saved {name.upper()} model to {path}")

if __name__ == "__main__":
    train_and_save()
    print("\n📦 AI SYSTEM: Multi-modal models are now REAL and operational.")
